fixed settings
